Remove restrictions on 0-width glyphs
The code currently prevents 0-width glyphs from being added to the PDF.  I'm not sure of the reason for this restriction, but there are cases when this needs to be allowed.

Our specific use case comes from a vietnamese string, which contains characters with various accents.  One example is the character that looks like this: https://www.compart.com/en/unicode/U+1EDD .  This character is built from two unicode characters: 0x01A1 and 0x0300.  The first character is the "o with horn" (https://www.compart.com/en/unicode/U+01A1).  The second character is the accent (https://www.compart.com/en/unicode/U+0300).  That second character is one of the "Combining Diacritical Marks" (https://www.compart.com/en/unicode/block/U+0300) - it's intended to "overlay" the previous character without taking up any space itself.

The current code omits that second character because it has no width.  This change removes that restriction.  I'm not sure why that restriction was there in the first place - my hope is that it was an optimization.

*** NOTE - this has been added to a forked copy of SVG-to-PDFKit, but was not added to its "master" branch.  The SVG-to-PDFKit we've been using was from npm version 0.1.8, which was published something like 4 years ago.  The version on master looks somewhat different, and the fix didn't work there.  So this is effectively freezing us at 0.1.8 and not picking up whatever changes have been happening since then (although I'm not sure what plans the author has for publishing those changes).

So the commit on "master" when I forked is now sitting on branch "master-2023-10-16" (for posterity), and this fix, which will be the new "master" is forked off of commit `be2b7bbe4cee36f15a31785772bc8291d46e8a7b`.  The comment for that commit is "Publish v0.1.8", so I'm assuming that's basically the same as what we've been using.